backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
redis==5.0.1
httpx==0.25.2
paypalcheckoutsdk==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
prometheus-client==0.19.0
sentry-sdk==1.39.1